Move constants.js out of utils
