+++ controllers/Pokemon.js
@@ -9,7 +9,7 @@ const {
   httpServerError,
   resOpSuccess,
   resOpFailure,
-} = require('../utils/constants');
+} = require('../constants');
 
 const pokemonController = {
   mongoGetAll: async (req, res) => {
